move events related tests to a separate file
